feat: move components to their route features
