+++ src/app/(site)/_components/album-card.tsx
@@ -5,7 +5,7 @@ import { BiDisc } from 'react-icons/bi';
 import useLoadImage from '~/hooks/useLoadImage';
 import { Album } from '~/types/types';
 import { formatDate } from '~/utils/format-date';
-import PlayButton from './PlayButton';
+import PlayButton from './play-button';
 
 interface AlbumCardProps {
   album: Album;
