+++ src/app/(site)/page.tsx
@@ -1,7 +1,7 @@
 import Header from '~/components/Header';
-import ListItem from '~/components/ListItem';
 import { getAlbums } from '~/server/actions/album/get-albums';
 import PageContent from './_components/PageContent';
+import ListItem from './_components/list-item';
 
 export const revalidate = 0;
